Move info dot next to file upload button
