+++ enterprise/frontend/src/metabase-enterprise/whitelabel/index.js
@@ -29,6 +29,7 @@ import { HelpLinkSettings } from "./components/HelpLinkSettings";
 import { IllustrationDescription } from "./components/IllustrationDescription";
 import { IllustrationWidget } from "./components/IllustrationWidget";
 import { ImageUpload } from "./components/ImageUpload";
+import { ImageUploadInfoDot } from "./components/ImageUploadInfoDot";
 import { LandingPageWidget } from "./components/LandingPageWidget";
 import LogoIcon from "./components/LogoIcon";
 import { MetabotToggleWidget } from "./components/MetabotToggleWidget";
@@ -222,6 +223,7 @@ if (hasPremiumFeature("whitelabel")) {
               customIllustrationSetting: "login-page-illustration-custom",
               errorMessageContainerId:
                 "login-page-illustration-error-container",
+              imageUploadInfoDot: <ImageUploadInfoDot type="login-page" />,
             },
           },
           {
@@ -242,6 +244,7 @@ if (hasPremiumFeature("whitelabel")) {
               customIllustrationSetting: "landing-page-illustration-custom",
               errorMessageContainerId:
                 "landing-page-illustration-error-container",
+              imageUploadInfoDot: <ImageUploadInfoDot type="landing-page" />,
             },
           },
           {
@@ -263,6 +266,9 @@ if (hasPremiumFeature("whitelabel")) {
                 "no-question-results-illustration-custom",
               errorMessageContainerId:
                 "no-question-results-illustration-error-container",
+              imageUploadInfoDot: (
+                <ImageUploadInfoDot type="no-question-results" />
+              ),
             },
           },
           {
@@ -284,6 +290,9 @@ if (hasPremiumFeature("whitelabel")) {
                 "no-search-results-illustration-custom",
               errorMessageContainerId:
                 "no-search-results-illustration-error-container",
+              imageUploadInfoDot: (
+                <ImageUploadInfoDot type="no-search-results" />
+              ),
             },
           },
         ],
